Resources objects created in sampleScene
Objects for each generable resource were created in sample scene and were added an script for each too.
ResourcesManager script was also added, wih a function that allows us to ask the XmlManager to increase the number of certain resource.
There is a problem in the file read though, solve it later.