[IMPORTANT] Coded the TempFile
Is quite possibly that all places where a fake index was placed, this index was replaced for a "real" index that is obtained from a temporal file that must exist only in runtime and must be filled with the gameIndex when a game is created or loaded.
This file, for testing purposes, will not be deleted on exit.
If want to access to a certain game file, you can either go to GameSelection interface and load it, or go and modify the TempFile directly.